ArrayInterface trait specialization
These changes are needed to support TileBuffer in LoopVectorization.
It contains one breaking change: formerly, `parent(tb::TileBuffer)`
returned the original backing array, whereas now it returns the `view`.
This is more consistent with the fact that the TileBuffer is
equivalent to the view in terms of indexing behavior.

This sets Julia 1.2 as lowest version in CI, as that is the lowest
version of Julia supported by ArrayInterface. For the tests, the
`@turbo` rename in LoopVectorization is only supported on Julia 1.5+.